lesson 12 nlp

+++ dl2/scripts/clean_nb_output.py
@@ -20,12 +20,13 @@ if __name__ == "__main__":
     for f in os.listdir():
         if os.path.isdir(f):
             continue
-        if f != 'clean_nb_output.py'and '_clean' not in f:
-            try:
-                pre, post = f.split('.')
-                newfile = pre+'_clean.py'
-                removeComments(f,newfile)
-                os.remove(f)
-                os.rename(newfile,f)
-            except Exception as e:
-                print(f'Error with {f}: {e}')
+        if f.endswith('.py'):
+            if f != 'clean_nb_output.py' and '_clean' not in f:
+                try:
+                    pre, post = f.split('.')
+                    newfile = pre+'_clean.py'
+                    removeComments(f,newfile)
+                    os.remove(f)
+                    os.rename(newfile,f)
+                except Exception as e:
+                    print(f'Error with {f}: {e}')
